requirements-dev.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.1.7
django-admin-interface==0.16.0
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
getch==1.0
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
django-environ==0.4.5
whitenoise==5.3.0
cloudinary==1.26.0
django-cloudinary-storage==0.3.0